refactor(LottoService): 로또 번호 긁는 로직을 LottoRepository로 이동
