+++ src/main/java/lotto/domain/lotto/LottoRepository.java
@@ -4,7 +4,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import lotto.domain.WinningLotto;
 import lotto.domain.lottomachine.LottoMachine;
+import lotto.domain.rating.Rating;
+import lotto.domain.rating.RatingInfo;
 
 public class LottoRepository {
 
@@ -18,6 +21,16 @@ public class LottoRepository {
 
     public void generateLottoByManual(final Lotto lotto) {
         lottos.add(lotto);
+    }
+
+    public RatingInfo scratchLotto(final WinningLotto winningLotto) {
+        RatingInfo ratingInfo = new RatingInfo();
+        for (Lotto lotto : lottos) {
+            int match = winningLotto.compareLottoNumber(lotto);
+            boolean hasBonusBall = winningLotto.compareBonusBall(lotto);
+            ratingInfo.update(Rating.getRating(match, hasBonusBall));
+        }
+        return ratingInfo;
     }
 
     public List<Lotto> toList() {
