+++ src/test/java/lotto/domain/lotto/LottoRepositoryTest.java
@@ -4,7 +4,14 @@ import static org.assertj.core.api.Assertions.*;
 
 import java.util.Arrays;
 import java.util.List;
+import lotto.domain.LottoService;
+import lotto.domain.Money;
+import lotto.domain.Ticket;
+import lotto.domain.WinningLotto;
+import lotto.domain.lottomachine.LottoMachine;
 import lotto.domain.lottomachine.RandomLottoMachine;
+import lotto.domain.rating.Rating;
+import lotto.domain.rating.RatingInfo;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -34,6 +41,24 @@ public class LottoRepositoryTest {
         LottoRepository lottoRepository = new LottoRepository();
         lottoRepository.generateLottoByManual(Lotto.from(Arrays.asList(1, 2, 3, 4, 5, 6)));
         assertThat(lottoRepository.toList().get(0).getNumbers()).isEqualTo(expected);
+    }
+
+    @Test
+    @DisplayName("로또 긁은 내역 확인")
+    void scratchLottoCheck() {
+        LottoService lottoService = new LottoService();
+        LottoRepository lottoRepository = new LottoRepository();
+        Ticket ticket = new Ticket(new Money(2000));
+
+        lottoService
+            .getLotto(lottoRepository, () -> Arrays.asList(1, 2, 3, 4, 5, 6), ticket.getCount());
+
+        WinningLotto winningLotto = new WinningLotto(
+            Lotto.from(Arrays.asList(1, 2, 3, 4, 5, 6)),
+            LottoNumber.from(7));
+        RatingInfo ratingInfo = lottoRepository.scratchLotto(winningLotto);
+
+        assertThat(ratingInfo.get(Rating.FIRST)).isEqualTo(2);
     }
 
 }
